+++ test/samples-common/construct-float.js
@@ -6,7 +6,6 @@ var expected = {
   canonical: 685230.15,
   exponential: 685230.15,
   fixed: 685230.15,
-  sexagesimal: 685230.15,
   'negative infinity': Number.NEGATIVE_INFINITY,
   'not a number': NaN
 };
@@ -18,7 +17,6 @@ function testHandler(actual) {
   assert.strictEqual(actual['canonical'],         expected['canonical']);
   assert.strictEqual(actual['exponential'],       expected['exponential']);
   assert.strictEqual(actual['fixed'],             expected['fixed']);
-  assert.strictEqual(actual['sexagesimal'],       expected['sexagesimal']);
   assert.strictEqual(actual['negative infinity'], expected['negative infinity']);
 
   assert(Number.isNaN(actual['not a number']));