+++ test/samples-common/construct-int.js
@@ -5,6 +5,5 @@ module.exports = {
   decimal: 685230,
   octal: 685230,
   hexadecimal: 685230,
-  binary: 685230,
-  sexagesimal: 685230
+  binary: 685230
 };